Modified to pass file descriptor instead of path
As an experiment, the file descriptor was passed on rather than
path to the tty. Refer
http://developer.android.com/reference/android/hardware/usb/UsbDeviceConnection.html#getFileDescriptor()
for documentation regarding file descriptor to be transferred.
The idea is that Java side of the project will obtain permission for
usage of USB device and pass it on to the native side. Using this
file descriptor, the libdivecomputer is expected to import dives.

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>