Revert to passing name rather than fd
As experiment 2 failed with the Error ENOTTY : Not a typewriter
we would revert to the original mechanism.
The reason behind this failure is explained by Jef.
The file descriptor given by android is actually the file
descriptor of the USB device not the tty built over it.
Hence it would not be possible to use libdivecomputer with this
file descriptor.
But maybe libusb can utilise this?

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>