Initial version serial_ftdi.c
This is based on serial_posix script.
There are a number of shortcomings.
Biggest of them is macro definition of VID and PID. Rather than this,
maintain a list of VID and PID of accepted FTDI devices.
Unimplemented functions like get_transmitted, send_break and set_break
return -1 on call.

For device_open(), vid pid would be better arguments than the path as
path is not too helpful for libftdi and libusb.

serial_enumerate gives back the path of all the devices present on all
hubs on android. Basically, all the files present in all the subfolders
of /dev/bus/usb.
A better implementation would be a set of ftdi devices attached to the
android. This could be done either using jni by enumerate_device in
android.cpp or using libusb and passing vid pid as an argument.

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>