Check for libftdi, include it in Required.private
Added check for libftdi during configure. This is necessary as libftdi would be
needed for communication on android. Suitable modifications are also made to
enable cross building for android platform.

If you are building for android, you would need libftdi package. Hence, add this
in Required.private so that it may be included automatically while cross
compilation.

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>