Finally a working version. Try again on no data.
ftdi_read_data returns 0 if no data is received from
the device. The existing code understood it as timeout errror.
Removing this check might end up in an infinite loop. But it works
in OSTC3. This should be checked further.
Cross compilation of libdivecomputer on android and running it
gives the results same as the ones on pc.

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>