Check multiple pids while opening device
libftdi returns -3 when it could not find any device corresponding to
the pid and vid given. (Different from -3 of libusb which means you do
not have permissions to access the device)
Using this value, check among a list of accepted pids. VID is that of
ftdi ( file is called serial_ftdi :)
There is a problem when multiple devices are connected. It uses the
first one it can find in the list. This might not be correct always.

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>